Shared memory between Buildsome and fshook to reduce chatter
This commit implements shared memory between Buildsome and FSHooks,
underwhich a hashtable is maintained, where Buildsome writes to
the hashtable and the FSHooks only read from it.

The purpose of this hashtable is to tell the FSHooks which files
were already seen completion of build handlers by Buildsome, and
so to prevent stalling over their Unix pipe when communicating
these file accesses to Buildsome. Thankfully, Buildsome already
implements no-wait file accesses - for out-of-tree access - so
this leverages over that.

The hashtable uses the SHA1 of the filenames accessed as keys,
and gets expanded to allow it to grow. The clients may
occasionally but gracefully fail when reading it due to races,
but it shouldn't affect the amortized saving in run-time. I
also linked the same C code both to Buildsome and to the FSHook
and used inline-c in order to create FFI bindings for it.

Further work can made for the accumulation of file access
notifications by using a per-target creation of shared memory,
to be written by the FSHook and be read by Buildsome upon
its termination. This would replace the many no-wait file access
notifications and would allow buildsome to batch-process these
file accesses at target execution termination.